requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.7
botocore==1.29.7
charset-normalizer==2.1.1
colorama==0.4.6
colorlog==6.7.0
frozenlist==1.3.3
hikari==2.0.0.dev112
hikari-lightbulb==2.2.5
hikari-miru==1.1.2
idna==3.4
jmespath==1.0.1
multidict==6.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
yarl==1.8.1
